pyproject.toml
[tool.poetry]
name = "git-1"
version = "0.1.0"
description = ""
authors = ["Spivak Matvii <spivak.matvii@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
passlib = "^1.7.4"
bcrypt = "^4.0.1"
pyjwt = "^2.8.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
flake8 = "^6.1.0"
black = "^23.11.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
